// npm-packages/brave-sdk/src/types/credit-bureau.ts
export interface IAddress {
  street: IStreet;
  location: ILocation;
  dateReported?: string;
  order?: string | number;
}

export interface IStreet extends ITUUnparsed {
  number: number | string;
  name: string;
}

export interface ILocation extends ITUUnparsed {
  city?: string;
  state?: string;
  zipCode?: number | string;
  zipExt?: number | string;
}
// ...
export interface ICreditBureauSubscriber {
  industryCode?: string;
  memberCode?: string;
  name: ITUUnparsed;
  address: IAddress;
  phone: IPhone;
}

export interface IPhone extends ITUUnparsed {
  number: IPhoneNumber;
}

export interface IPhoneNumber extends ITUUnparsed {
  areaCode: number | string;
  exchange: number | string;
  suffix: number | string;
}